refactoring head to be relative to body
