+++ client/components/aframe/bodyParts/Body.js
@@ -1,4 +1,5 @@
 import React from 'react'
+import Head from './Head'
 
 const Body = props => {
   return (
@@ -12,9 +13,12 @@ const Body = props => {
       position="0.024 0.324 -1.501"
       geometry=""
     >
+      <Head />
       {props.animation}
     </a-box>
   )
 }
 
 export default Body
+
+// animation={head && head.map(animation => animation)}
